requirements.txt
Flask
mysql-connector-python
tensorflow
numpy
Pillow
gunicorn
python-dotenv
scipy